chore: fix storybook
