--- .storybook/preview.ts
@@ -1,19 +1,0 @@
-import type { Preview } from '@storybook/react';
-import '../src/styles/globals.css';
-
-const preview: Preview = {
-  parameters: {
-    actions: { argTypesRegex: '^on[A-Z].*' },
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/,
-      },
-    },
-    nextjs: {
-      appDirectory: true,
-    },
-  },
-};
-
-export default preview;
